Revert "Merge branch '37-create-quiz-ui' of https://github.com/DIT113-V22/group-01 into 39-update-the-quiz-state"
This reverts commit 6e07dc219dbf1af6fe54900d58a14a6547e8af99, reversing
changes made to 73083c79580d8bb74f99023f24cc88dd5df3e44a.
