+++ SmartCarMQTTApp/app/src/main/java/com/example/smartcarmqttapp/state/QuizState.java
@@ -61,21 +61,21 @@ public class QuizState {
             // Answer is incorrect. Add question to review
              questions.get(currentPointer).setNeedsReview(1);
         }
-        //currentAnswers.add(answer); // adds answer to list of current answers
+        currentAnswers.add(answer); // adds answer to list of current answers
         incrementCurrentPointer(); // increases question number
     }
 
-    public Question getCurrentQuestion(int currentPointer) {
+    public Question getCurrentQuestion() {
         return this.questions.get(currentPointer);
     }
 
-    public void incrementScore() {
+    private void incrementScore() {
         this.score++;
         // Made method separate in case UI actions needed
     }
 
     // Moves one question up (we don't need moving down)
-    public void incrementCurrentPointer() {
+    private void incrementCurrentPointer() {
         this.currentPointer++;
         if (this.currentPointer + 1 == this.questions.size())
             finishQuiz();
