--- SmartCarMQTTApp/app/src/main/res/drawable/ic_baseline_follow_the_signs_24.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M9.5,5.5c1.1,0 2,-0.9 2,-2s-0.9,-2 -2,-2s-2,0.9 -2,2S8.4,5.5 9.5,5.5zM5.75,8.9L3,23h2.1l1.75,-8L9,17v6h2v-7.55L8.95,13.4l0.6,-3C10.85,12 12.8,13 15,13v-2c-1.85,0 -3.45,-1 -4.35,-2.45L9.7,6.95C9.35,6.35 8.7,6 8,6C7.75,6 7.5,6.05 7.25,6.15L2,8.3V13h2V9.65L5.75,8.9M13,2v7h3.75v14h1.5V9H22V2H13zM18.01,8V6.25H14.5v-1.5h3.51V3l2.49,2.5L18.01,8z"/>
-</vector>
