--- SmartCarMQTTApp/app/src/main/res/drawable/progressbar_circle.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-        <shape
-            android:shape="ring"
-            android:thicknessRatio="14"
-            android:useLevel="false">
-            <solid android:color="#DDD" />
-        </shape>
-    </item>
-
-    <item>
-        <shape
-            android:shape="ring"
-            android:thicknessRatio="14"
-            android:useLevel="true">
-            <solid android:color="@color/cadmiumGreen"/>
-        </shape>
-    </item>
-
-</layer-list>
