+++ SmartCarMQTTApp/app/src/main/res/layout/activity_about_us.xml
@@ -4,12 +4,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PracticeTheoryActivity">
+    tools:context=".AboutUsActivity">
 
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="This is result screen for quiz theory"
+        android:text="This is about us"
         android:textSize="30dp"
         android:layout_centerInParent="true"
         app:layout_constraintBottom_toBottomOf="parent"
